+++ engine/src/parser/expression_utils.cpp
@@ -1460,3 +1460,39 @@ std::string modify_multi_column_count_expression(std::string expression, std::ve
 
 	return expression;
 }
+
+std::string get_current_date_or_timestamp(std::string expression, blazingdb::manager::Context * context) {
+    // We want `CURRENT_TIME` holds the same value as `CURRENT_TIMESTAMP`
+	if (expression.find("CURRENT_TIME") != expression.npos) {
+		expression = StringUtil::replace(expression, "CURRENT_TIME", "CURRENT_TIMESTAMP");
+	}
+
+	std::size_t date_pos = expression.find("CURRENT_DATE");
+	std::size_t timestamp_pos = expression.find("CURRENT_TIMESTAMP");
+
+	if (date_pos == expression.npos && timestamp_pos == expression.npos) {
+		return expression;
+	}
+
+    // CURRENT_TIMESTAMP will return a `ms` format
+	std::string	timestamp_str = context->getCurrentTimestamp().substr(0, 23);
+    std::string str_to_replace = "CURRENT_TIMESTAMP";
+
+	// In case CURRENT_DATE we want only the date value
+	if (date_pos != expression.npos) {
+		str_to_replace = "CURRENT_DATE";
+        timestamp_str = timestamp_str.substr(0, 10);
+	}
+
+	return StringUtil::replace(expression, str_to_replace, timestamp_str);
+}
+
+std::string preprocess_expression_for_evaluation(std::string expression, blazingdb::manager::Context * context, std::vector<cudf::data_type> schema) {
+	expression = fill_minus_op_with_zero(expression);
+	expression = convert_concat_expression_into_multiple_binary_concat_ops(expression);
+	expression = get_current_date_or_timestamp(expression, context);
+	expression = convert_ms_to_ns_units(expression);
+	expression = reinterpret_timestamp(expression, schema);
+	expression = apply_interval_conversion(expression, schema);
+  return expression;
+}