+++ CHANGELOG.md
@@ -6,9 +6,12 @@
 ## Improvements
 - #1571 Update ucx-py versions to 0.21
 - #1554 return ok for filesystems
+- #1572 Setting up default value for max_bytes_chunk_read to 256 MB
+
 
 ## Bug Fixes
-- #1570 Fix build due to changes in rmm device buffer 
+- #1570 Fix build due to changes in rmm device buffer
+- #1576 Fix `CC`/`CXX` variables in CI
 
 
 # BlazingSQL 21.06.00 (June 10th, 2021)